server/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.66.3
outcome==1.3.0.post0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
selenium==4.29.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.39
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
